rag-backend/requirements.txt
sentence-transformers
torch
qdrant-client
python-dotenv
fastapi
uvicorn
pydantic
